% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_dataset.R
\name{clim_dataset}
\alias{clim_dataset}
\title{Create a climate dataset object}
\usage{
clim_dataset(
  data = NULL,
  vars = list(RAIN = "RAIN", TMAX = "TMAX", TMIN = "TMIN", TMEAN = "TMEAN"),
  time_vars = list(DATE = "DATE", YEAR = "YEAR", MONTH = "MONTH", DAY = "DAY"),
  coords = list(LONG = "LONG", LAT = "LAT"),
  t_step = "daily",
  verbose = TRUE
)
}
\arguments{
\item{date}{is the currente date to be converted}

\item{format}{this parameter controle the format of the date}
}
\value{
result is a julian day
}
\description{
clim_dataset creates a climate dataset object structure using s3 system.
}
\examples{

clim_dataset(data = my_dataset) ## will return a clim dataset object

}
