% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateoflastfull.r
\name{dateoflastefull}
\alias{dateoflastefull}
\title{Compute the last dates of soil full capacity}
\usage{
dateoflastefull(
  x,
  from = 1,
  to = length(x),
  soil_cap = 70,
  replaceNA = TRUE,
  NAValues = 0
)
}
\arguments{
\item{x}{the rainfall vector}

\item{from}{from which element of x to start}

\item{to}{last element for dry spell computing}

\item{soil_cap}{maximum capacity of water storage of th soil (default to 70.0 in Burkina Faso)}

\item{replaceNA}{Replace missing values by a specified value defaulted to 0.0 mm}

\item{threshold}{the default value of rain quantity to be counted as a rainy day}

\item{NAvalues}{missing data replacement values}
}
\value{
result is a julian day (integer)
}
\description{
dateoflastefull compute the laste date of the soil full capacity starting from the begin of vector x
}
\examples{

dateoflastefull(x) ## will return the last date when we have the maximum capacity of the soil

}
