% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy.R
\name{doy}
\alias{doy}
\title{Extract the number of day in a year or julian date}
\usage{
doy(date, format = "\%Y/\%m/\%d")
}
\arguments{
\item{date}{is the currente date to be converted}

\item{format}{this parameter controle the format of the date}
}
\value{
result is a julian day
}
\description{
doy extract the julian date from the date
}
\examples{

doy( date = as.Date('1990/05/23')) ## will return the julian day corresponding to the date '1990/05/23'

}
