% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_spell.r
\name{dry_spell}
\alias{dry_spell}
\title{Compute the dry spell in rainy  season}
\usage{
dry_spell(
  x,
  threshold = 0.85,
  replaceNA = TRUE,
  NAvalues = 0,
  MaxNAPercent = 0.25,
  return_max = TRUE
)
}
\arguments{
\item{x}{the rainfall vector}

\item{threshold}{the default value of rain quantity to be counted as a rainy day}

\item{replaceNA}{Replace missing values by a specified value defaulted to 0.0 mm}

\item{NAvalues}{missing data replacement values}

\item{MaxNAPercent}{maximum percentage of missing data}

\item{return_max}{return maximum or detail values}
}
\value{
result is a consecutive number of dry days
}
\description{
dry_spell function compute the maximul dry spell or all dry spells in rainfall data
}
\examples{

dry_spell(x) ## will return the maximum dry spell

}
