% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_spell_2.r
\name{dry_spell_2}
\alias{dry_spell_2}
\title{Compute the dry spell in rainy  season}
\usage{
dry_spell_2(
  x,
  threshold = 0.85,
  NAValues = 0,
  from = 1,
  to = length(x),
  IncludeNA = TRUE,
  DropNA = FALSE,
  NaMaxPercent = 0.25
)
}
\arguments{
\item{x}{the rainfall vector}

\item{threshold}{the default value of rain quantity to be counted as a rainy day}

\item{from}{from which element of x to start}

\item{to}{last element for dry spell computing}

\item{IncludeNA}{Include missing value in the output result}

\item{DropNA}{drop missing values before computing}

\item{replaceNA}{Replace missing values by a specified value defaulted to 0.0 mm}

\item{NAvalues}{missing data replacement values}

\item{MaxNAPercent}{maximum percentage of missing data}
}
\value{
result is a consecutive number of dry days
}
\description{
dry_spell function compute the maximul dry spell or all dry spells in rainfall data
}
\examples{

dry_spell_2(x) ## will return the maximum dry spell

}
