% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dspell_end.r
\name{Dspell_end}
\alias{Dspell_end}
\title{Compute the dry spell at the end of the season}
\usage{
Dspell_end(
  x,
  from = dt_start(x),
  to = end_season(x),
  Fromstart = 50,
  threshold = 0.85,
  Fromlastrain = TRUE
)
}
\arguments{
\item{x}{the rainfall vector}

\item{from}{from which element of x to start}

\item{to}{laste date for the computation it usually correspond to the end of the season.}

\item{Fromstart}{This parameter control how many days from the start of the season to look for end spell}

\item{threshold}{rain threshold. Minimum rain quantity to be considered as a beneficial rain}

\item{Fromlastrain}{this parameter controle if to start from a rainy day or not}
}
\value{
result is a consecutive number of dry days
}
\description{
Dspell_end function compute the dry spell at the end of the rainy season.
}
\examples{

Dspell_end(x) # will return the end of season dry spell x is rainfall vector

}
