% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dspell_start.r
\name{Dspell_start}
\alias{Dspell_start}
\title{Compute the dry spell at the begining of the rainy season}
\usage{
Dspell_start(
  x,
  from = dt_start(x),
  period = 50,
  threshold = 0.85,
  Tolastrain = TRUE
)
}
\arguments{
\item{x}{the rainfall vector}

\item{from}{from which element of x to start}

\item{period}{period of time in days from rain start to bloom}

\item{threshold}{the default value of rain quantity to be counted as a rainy day}

\item{Tolastrain}{this parameter controls the endpoint of the dry spell period}
}
\value{
result is a consecutive number of dry days
}
\description{
Dspell_start function compute the dry spell at the begining of the rainy season.
}
\examples{

Dspell_start(x) # will return the start of season dry spell where x is rainfall vector

}
