% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_start.r
\name{dt_start}
\alias{dt_start}
\title{Compute the start of rainy season}
\usage{
dt_start(
  x,
  early = 91,
  limit = length(x),
  SumVdays = 20,
  n_Vdays = 3,
  maxDrySpell_aft = 20,
  onMaxdays = 30
)
}
\arguments{
\item{x}{the rainfall vector}

\item{limit}{the maximum length of x or the latest not included date}

\item{SumVdays}{the total rainfall over V (n_Vdays) consicutive days for starting condition}

\item{n_Vdays}{numbers of consecutive days to compute SumVdays on}

\item{maxDrySpell_aft}{maximum of dry spell after SumVdays condition realized}

\item{onMaxdays}{maximum of consecutive days for maxDrySpell computation}
}
\value{
result in percent of NA number
}
\description{
dt_start function compute the start of the rainy season in julian days using Sivakoumar like conditions
}
\examples{

dt_start(x) ## return the begin of the rainy season

}
