% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_season.r
\name{end_season}
\alias{end_season}
\title{Compute the end of the rainy season}
\usage{
end_season(
  x,
  from = 245,
  to = length(x),
  replaceNA = FALSE,
  NAValues = 0,
  soil_cap = 70,
  UseLastFull = FALSE,
  daily_etp = 5
)
}
\arguments{
\item{x}{the rainfall vector}

\item{from}{from which element of x to start}

\item{to}{last element for dry spell computing}

\item{replaceNA}{Replace missing values by a specified value defaulted to 0.0 mm}

\item{soil_cap}{maximum capacity of water storage of th soil (default to 70.0 in Burkina Faso)}

\item{UseLastFull}{use late full storage date to starte calculation}

\item{threshold}{the default value of rain quantity to be counted as a rainy day}

\item{NAvalues}{missing data replacement values}

\item{MaxNAPercent}{maximum percentage of missing data}
}
\value{
result is a consecutive number of dry days
}
\description{
dry_spell function compute the end of the rainy season in julian days using water balance methode.
}
\examples{

end_season(x) ## return the end of the rainy season

}
