% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leapyear.R
\name{is_leapyear}
\alias{is_leapyear}
\title{Test if a given year is leap year or note}
\usage{
is_leapyear(year = 2021)
}
\arguments{
\item{year}{this parameter is the year of the date}
}
\value{
result is a boolean
}
\description{
This function return True if a year is a leap year and False otherwise.
}
\examples{

is_leapyear(2020) ## TRUE
is_leapyear(2021) ## FALSE

}
