% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_todate.r
\name{julian_todate}
\alias{julian_todate}
\title{Convert julian days to normal days}
\usage{
julian_todate(julian, year = NULL, leap = TRUE, format = "\%yyyy/\%mm/\%dd")
}
\arguments{
\item{julian}{is the julian date (number of days since the degining of the year)}

\item{year}{the year of the date (julian dates dont carry year component) default null}

\item{leap}{This specify if year is leap year or not}

\item{format}{this parameter controle the format of the date}
}
\value{
result is a date or a date like format string (%dd/%mm) if year not specify
}
\description{
julian_todate function convert julian dates to normal dates
}
\examples{

julian_1 <- julian_todate(178)
julian_2 <-  julian_todate(-20) ## will return NA because day must be positive
julian_3 <- julian_todate(390) ## will return NA because days must be less than 366

}
