% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxDrySpell.r
\name{maxDrySpell}
\alias{maxDrySpell}
\title{Compute the maximum dry spell in a rainfall vector}
\usage{
maxDrySpell(x, threshold = 0.85, from = 1, to = length(x))
}
\arguments{
\item{x}{A numeric vector of rainfall}

\item{threshold}{the daily rainfall threshold}

\item{from}{start computation from (default to 1)}

\item{to}{end of computation}
}
\value{
result is an integer giving the maximum dry spell
}
\description{
This function compute the maximum dry spell in a rainfall vector
}
\examples{


maxDrySpell(x) ## will return the maximum dry spell in a year

}
