% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NApercent.r
\name{NAPercent}
\alias{NAPercent}
\title{Compute the missing data percent}
\usage{
NAPercent(x)
}
\arguments{
\item{x}{the rainfall vector}
}
\value{
result in percent of NA number
}
\description{
This function compute the percentage of NA values in a numeric vector (rainfall data)
}
\examples{
Np <- NAPercent( c(0.0, 0.4, 45.2, NA, 21, NA, NA, NA))  # should give 50 (\%)
Np <- NAPercent( c(rnorm(75), rep(NA, 25)))  # should give 25 (\%)

}
