% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_at.r
\name{total_at}
\alias{total_at}
\title{Compute the exact days or step a certain total quantity is obtain}
\usage{
total_at(x, total = 20)
}
\arguments{
\item{total}{the total rainfall to have}
}
\value{
result the step
}
\description{
dt_start function compute the start of the rainy season in julian days using Sivakoumar like conditions
}
\examples{

total_at(x) ## will return the last days a certain quantity of rain is obtain

}
